Use int to count errors and warnings in FormattingDocumentedHandler
To avoid potential overflow.

+++ community/consistency-check/src/main/java/org/neo4j/consistency/report/ConsistencyReporter.java
@@ -178,8 +178,8 @@ public class ConsistencyReporter implements ConsistencyReport.Reporter
     {
         private final InconsistencyReport report;
         private final RecordType type;
-        private short errors;
-        private short warnings;
+        private int errors;
+        private int warnings;
 
         FormattingDocumentedHandler( InconsistencyReport report, RecordType type )
         {
